Refactor: Update GenreSelect component and modify modals
This commit includes the following changes:

Updated the GenreSelect component to use a new list of genres and a new way to display them.
Modified the EditModal and UploadModal components to directly connect the onGenreChange property of the GenreSelect component to the setSelectedGenres function.
Modified the style of the button in the PlaylistModal component.
Added marginBottom to the UserCard component in the Sidebar component.